migrates stream handling logic
